Renamed public types
